ECPRESOURCECHEF-26
Added _DeleteNode procedure to teardown the Chef API client and node. This procedure should be invoked by dynamic environments before teardown of the resource can be invoked.

o _DeleteNode procedure accepts one parameter: resource name, which can be a resource instance or a resource pool that was previously provisioned by dynamic environments.
o A single step will be broadcast to all the resource(s) (all the resources in the resource pool are assumed to be dynamic resources) to delete its chef API client and chef node.
o Updated the ec_configurationmanagement_plugin property sheet to include 'teardown' as a new operation that maps to this procedure. The wiki doc will also be updated with these changes.

[git-p4: depot-paths = "//depot/nimbus/integrations/EC-Chef/main/": change = 99349]