CEV-6705 Updated the _DeleteNode procedure to not fail in case of missing properties on the resource being torn down or in case of a general failure to delete the chef node or client.  It is possible that converge may not have completed successfully and the _DeleteNode procedure is being called as part of the cleanup process in which case we would have a resource which does not have the properties setup required for chef or have a partially setup chef client or node.
There are two kinds of errors that _DeleteNode would have failed on earlier. One is if there were missing properties on the resource whose chef node and client are being deleted. These properties would be missing in case the converge process had failed earlier so in this condition, now the _DeleteProcedure will simply complete after logging a message. The other error condition in which the _DeleteNode would have failed is if there were missing chef files or the deletion command itself failed. In this case now, the _DeleteProcedure will complete after logging a warning.

[git-p4: depot-paths = "//depot/nimbus/integrations/EC-Chef/main/": change = 100169]