ECPRESOURCECHEF-44
Since this is a basic communication issue between the agent and the server, we can only have the agent log an error message if it fails to communicate with the server. The agent cannot stop the job since that also requires talking to the server. So, addressing the issue by invoking getServerStatus() in the procedure to test the agent->commanderServer communication. The call will log any error encountered while trying to communicate with the commander server.

[git-p4: depot-paths = "//depot/nimbus/integrations/EC-Chef/main/": change = 103334]